Alert box added

+++ components/FavoriteComponent.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
-import { View, FlatList } from 'react-native';
+import { View, FlatList,Text,Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import {connect} from 'react-redux';
 import {baseUrl} from '../shared/baseUrl';
 import {Loading} from './LoadingComponent';
-import SwipeOut from 'react-native-swipeout';
+import Swipeout from 'react-native-swipeout';
 import {deleteFavorite} from '../redux/ActionCreator';
 
 const mapStateToProps = state => {
@@ -37,7 +37,25 @@ class Favorites extends Component{
                 {
                     text: 'Delete',
                     type: 'Delete',
-                    onPress: ()=>this.props.deleteFavorite(item.id)
+                    onPress: ()=>{
+                        Alert.alert(
+                            'Delete Favorite',
+                            'Are you sure you wish to delete the favorite dish '+ item.name+' ?',
+                            [
+                                {
+                                    text:'Cancel',
+                                    onPress: ()=> console.log(item.name+' Not Deleted'),
+                                    style: 'cancel'
+                                },
+                                {
+                                    text:'Ok',
+                                    onPress: ()=> this.props.deleteFavorite(item.id),
+                                    
+                                }
+                            ],
+                            {cancelable: false}
+                        );
+                    }
                 }
             ];
